refactor dashboard into services and repositories
