+++ src/singular/dashboard/repositories/__init__.py
@@ -1,0 +1,1 @@
+"""Dashboard repositories for persistent data access."""
