+++ src/singular/dashboard/routes/__init__.py
@@ -1,0 +1,1 @@
+"""Dashboard route modules (migration scaffold)."""
