+++ src/singular/dashboard/services/__init__.py
@@ -1,0 +1,1 @@
+"""Dashboard service layer for business logic."""
